Ottimizzazioni per l'autenticazione e il login
