+++ HTML/partita.html
@@ -1,0 +1,11 @@
+<!DOCTYPE html>
+<html lang="it-IT">
+<head>
+    <meta charset="UTF-8">
+    <meta name="viewport" content="width=device-width, initial-scale=1.0">
+    <title>Partita</title>
+</head>
+<body>
+    <!--TODO: Questa potrebbe essere la pagina della partita del gioco con tutta l'implementazione di socket.io-->
+</body>
+</html>
